Proxy is now written in pure C
And is renamed from proxy to injector_helper
