+++ .gitignore
@@ -33,3 +33,4 @@
 *.su
 
 injector
+injector_helper
